requirements.txt
elasticsearch==7.16.0
itemadapter==0.4.0
PyYAML==6.0
Scrapy==2.5.1
scrapy_splash==0.8.0
Twisted~=21.7.0